+++ ur_robot_driver/test/integration_test_controller_switch.py
@@ -60,21 +60,17 @@ ALL_CONTROLLERS = [
 
 
 @pytest.mark.launch_test
-@launch_testing.parametrize(
-    "tf_prefix",
-    [""],
-    # [(""), ("my_ur_")],
-)
+@launch_testing.parametrize("tf_prefix", [(""), ("my_ur_")])
 def generate_test_description(tf_prefix):
     return generate_driver_test_description(tf_prefix=tf_prefix)
 
 
-class RobotDriverTest(unittest.TestCase):
+class ControllerSwitchTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         # Initialize the ROS context
         rclpy.init()
-        cls.node = Node("robot_driver_test")
+        cls.node = Node("controller_switching_test")
         time.sleep(1)
         cls.init_robot(cls)
 
@@ -207,11 +203,21 @@ class RobotDriverTest(unittest.TestCase):
                 ],
             ).ok
         )
+        # This got removed on 2025-10-29 due to a change in ros2_control
+        # See https://github.com/ros-controls/ros2_control/issues/2758 for details
+        # self.assertFalse(
+        # self._controller_manager_interface.switch_controller(
+        # strictness=SwitchController.Request.STRICT,
+        # activate_controllers=[
+        # "forward_position_controller",
+        # ],
+        # ).ok
+        # )
         self.assertFalse(
             self._controller_manager_interface.switch_controller(
                 strictness=SwitchController.Request.STRICT,
                 activate_controllers=[
-                    "forward_position_controller",
+                    "forward_velocity_controller",
                 ],
             ).ok
         )
@@ -219,7 +225,7 @@ class RobotDriverTest(unittest.TestCase):
             self._controller_manager_interface.switch_controller(
                 strictness=SwitchController.Request.STRICT,
                 activate_controllers=[
-                    "forward_velocity_controller",
+                    "forward_effort_controller",
                 ],
             ).ok
         )