+++ ur_calibration/CHANGELOG.rst
@@ -2,6 +2,18 @@
 Changelog for package ur_calibration
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+4.3.0 (2025-10-14)
+------------------
+
+4.2.0 (2025-09-16)
+------------------
+
+4.1.0 (2025-07-29)
+------------------
+* Use hpp headers from geometry2 (`#1467 <https://github.com/UniversalRobots/Universal_Robots_ROS2_Driver/issues/1467>`_)
+* Fix ur_calibration compilation on Windows (`#1400 <https://github.com/UniversalRobots/Universal_Robots_ROS2_Driver/issues/1400>`_)
+* Contributors: Felix Exner, Silvio Traversaro
+
 4.0.3 (2025-06-16)
 ------------------