+++ ur_moveit_config/CHANGELOG.rst
@@ -2,6 +2,19 @@
 Changelog for package ur_moveit_config
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+4.3.0 (2025-10-14)
+------------------
+* Add support for UR18 (`#1524 <https://github.com/UniversalRobots/Universal_Robots_ROS2_Driver/issues/1524>`_)
+* Contributors: Felix Exner
+
+4.2.0 (2025-09-16)
+------------------
+* Add support for launching UR8Long (`#1490 <https://github.com/UniversalRobots/Universal_Robots_ROS2_Driver/issues/1490>`_)
+* Contributors: Felix Exner
+
+4.1.0 (2025-07-29)
+------------------
+
 4.0.3 (2025-06-16)
 ------------------